refactor: decompose ArmyViewPage into focused hooks and handlers (#225)
